Custom CSS and JavaScript
Here is my custom CSS & Javascript on the website. The Javascript is used for the toggling on and off the packages and the FAQ. I was unable to post the whole WP website to GitHub, therefore I have just included the Customization in a new folder here. The Javascript is hidden on the Packages page and FAQ page on a shortcode row.